September, Day 16 Problem Type - Xor. Trie
Need to go through trie once more.

Not getting the adequate idea when to use it when not.

Links for refernce -
i.	Without trie - https://leetcode.com/problems/maximum-xor-of-two-numbers-in-an-array/discuss/849540/python-solution-with-detail-comments
ii.	https://leetcode.com/problems/maximum-xor-of-two-numbers-in-an-array/discuss/849775/Python-Simple-Solution-(beginners)-(video-%2B-co Xor. Trie
iii.	https://leetcode.com/problems/maximum-xor-of-two-numbers-in-an-array/discuss/849102/Python-O(n)-by-Trie-with-binary-notation-w-Comment(